@@ -11,29 +11,29 @@
 		"@lucia-auth/oauth": "^3.5.0",
 		"drizzle-orm": "^0.29.1",
 		"lucia": "^2.7.4",
-		"open-props": "1.6.13",
+		"open-props": "1.6.16",
 		"pretty-ms": "^8.0.0"
 	},
 	"devDependencies": {
 		"@melt-ui/pp": "0.1.4",
-		"@melt-ui/svelte": "0.63.1",
+		"@melt-ui/svelte": "0.65.0",
 		"@playwright/test": "^1.40.1",
 		"@sveltejs/adapter-auto": "^2.1.1",
 		"@sveltejs/adapter-netlify": "2.0.8",
 		"@sveltejs/enhanced-img": "0.1.5",
-		"@sveltejs/kit": "^1.27.6",
-		"@types/node": "20.10.1",
-		"@typescript-eslint/eslint-plugin": "^6.13.1",
-		"@typescript-eslint/parser": "^6.13.1",
+		"@sveltejs/kit": "^1.27.7",
+		"@types/node": "20.10.4",
+		"@typescript-eslint/eslint-plugin": "^6.13.2",
+		"@typescript-eslint/parser": "^6.13.2",
 		"autoprefixer": "10.4.16",
 		"dotenv": "^16.3.1",
 		"drizzle-kit": "^0.20.6",
-		"eslint": "^8.54.0",
-		"eslint-config-prettier": "^9.0.0",
+		"eslint": "^8.55.0",
+		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.35.1",
 		"lefthook": "1.5.5",
 		"npm-run-all": "4.1.5",
-		"postcss": "8.4.31",
+		"postcss": "8.4.32",
 		"postcss-jit-props": "1.0.14",
 		"prettier": "^3.1.0",
 		"prettier-plugin-svelte": "^3.1.2",
@@ -41,10 +41,10 @@
 		"svelte-check": "^3.6.2",
 		"svelte-sequential-preprocessor": "2.0.1",
 		"tslib": "^2.6.2",
-		"tsx": "^4.6.1",
-		"typescript": "^5.3.2",
+		"tsx": "^4.6.2",
+		"typescript": "^5.3.3",
 		"vite": "^4.5.0",
-		"vitest": "^0.34.6"
+		"vitest": "^1.0.2"
 	},
 	"private": true,
 	"scripts": {
