@@ -7,7 +7,7 @@
 	import TrackArtists from "./track-artists.svelte";
 	import VendorLinks from "./vendor-links.svelte";
 	import { getTrackLinks } from "$lib/utils/track";
-	import { onPlayBtnClick } from "$lib/utils/player";
+	import { playTrack } from "$lib/utils/player";
 
 	export let index: number;
 	export let track: AudioTrack;
@@ -35,7 +35,7 @@
 		<VendorLinks {links} />
 	</div>
 
-	<button class="playbtn" on:click={() => onPlayBtnClick(track.id)}>
+	<button class="playbtn" on:click={() => playTrack(track.id)}>
 		<Icon id="icon-play-btn" />
 		<span class="sr-only">Play</span>
 	</button>
